requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
pypdf
python-docx
python-dotenv
google-generativeai>=0.8.0